Avoid ajax errors by not using ajax
This UI necessitates rapid layout changes, and would need frequent ajax callbacks to work. However, we don't want to display an inline preview as part of this UI, so we can avoid calling back to the server for each change.

So this version of the UI implements a hidden textfield which facilitates the communication between the client and server. The JS and the server side manages the data so it is properly converted as needed. The layout is saved on form save, no temporary storage is used for changes in-between.